requirements.txt
paho-mqtt
requests
logging
json